project/requirements.txt
Flask
Flask-Cors
Pillow
Werkzeug
pycryptodome
colorama
